portfolio_read/src/requirements.txt
requests
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
automaton==3.2.0
autopage==0.5.2
bcrypt==4.1.2
bison==0.1.3
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cliff==4.5.0
cmd2==2.4.3
cryptography==42.0.1
debtcollector==2.5.0
decorator==5.1.1
distlib==0.3.8
dnspython==2.4.2
docker==7.0.0
dogpile.cache==1.3.0
eventlet==0.35.0
exceptiongroup==1.2.0
fastapi==0.109.0
fasteners==0.19
filelock==3.13.1
fixtures==4.1.0
futurist==2.4.1
greenlet==3.0.3
h11==0.14.0
idna==3.6
importlib-metadata==7.0.1
iso8601==2.1.0
Jinja2==3.1.3
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keystoneauth1==5.5.0
keystonemiddleware==10.5.0
kombu==5.3.5
logger==1.4
logutils==0.3.5
magnum==17.0.1
Mako==1.3.0
mangum==0.17.0
MarkupSafe==2.1.4
motor==3.3.2
msgpack==1.0.7
netaddr==0.10.1
netifaces==0.11.0
networkx==3.2.1
openstacksdk==2.1.0
os-client-config==2.1.0
os-service-types==1.7.0
osc-lib==3.0.0
oslo.cache==3.6.0
oslo.concurrency==5.3.0
oslo.config==9.3.0
oslo.context==5.3.0
oslo.db==14.1.0
oslo.i18n==6.2.0
oslo.log==5.4.0
oslo.messaging==14.6.0
oslo.metrics==0.7.0
oslo.middleware==6.0.0
oslo.policy==4.2.1
oslo.reports==3.2.0
oslo.serialization==5.3.0
oslo.service==3.3.0
oslo.upgradecheck==2.2.0
oslo.utils==7.0.0
oslo.versionedobjects==3.2.0
packaging==23.2
Paste==3.7.1
PasteDeploy==3.1.0
pbr==6.0.0
pecan==1.5.1
pip==23.3.2
pipenv==2023.11.15
platformdirs==4.1.0
prettytable==3.9.0
prometheus-client==0.19.0
psutil==5.9.8
pycadf==3.1.1
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydantic-settings==2.1.0
pydot==2.0.0
PyJWT==2.8.0
pymongo==4.6.1
pyOpenSSL==24.0.0
pyparsing==3.1.1
pyperclip==1.8.2
python-barbicanclient==5.6.1
python-cinderclient==9.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-glanceclient==4.4.0
python-heatclient==3.4.0
python-keystoneclient==5.3.0
python-neutronclient==11.1.0
python-novaclient==18.4.0
python-octaviaclient==3.6.0
python-openstackclient==6.4.0
python-swiftclient==4.4.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
repoze.lru==0.7
requests==2.31.0
requestsexceptions==1.4.0
rfc3986==2.0.0
Routes==2.5.1
rpds-py==0.17.1
setuptools==69.0.3
simplegeneric==0.8.1
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
statsd==4.0.1
stevedore==5.1.0
taskflow==5.4.0
tenacity==8.2.3
testresources==2.0.1
testscenarios==0.5.0
testtools==2.7.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.25.0
vine==5.1.0
virtualenv==20.25.0
warlock==2.0.1
wcwidth==0.2.13
WebOb==1.8.7
Werkzeug==3.0.1
wheel==0.42.0
wrapt==1.16.0
WSME==0.12.1
yappi==1.6.0
zipp==3.17.0